Refactor the DataTransfer objects
