--- DataTransferObject/Form/UserInterface.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace SumoCoders\FrameworkMultiUserBundle\DataTransferObject\Form;
-
-interface UserInterface
-{
-    /**
-     * @return string
-     */
-    public function getClass();
-}
