+++ DataTransferObject/UserDataTransferObject.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SumoCoders\FrameworkMultiUserBundle\DataTransferObject;
+
+use SumoCoders\FrameworkMultiUserBundle\User\User;
+
+interface UserDataTransferObject
+{
+    public static function fromUser(User $user);
+    public function getEntity();
+}
